refactor(query): split inspect helpers into inspect_util to stay under line limit
